requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.30,<1.0
pydantic-settings>=2.10,<3.0
python-dotenv>=1.1,<2.0
alembic>=1.13,<2.0
